Update prob-and-stat.Rmd
These changes edit Chapter 5 so that the notation used there mimics the notation used in Chapters 8, Section 9.2.2,  Section 17.3.6, and perhaps elsewhere (those were the sections where I saw the alternative notation used).  Specifically, chapter 5 has \alpha/2 as an input to a function, whereas Chapters 8, section 9.2.2, and section 17.3.6 use \alpha/2 as a subscript.  The actual notation used doesn't matter so much, but it should be consistent within a single text.  So I chose to edit Chapter 5, since that seems to be the less frequently used notation.